refactor: Move useDashboardList hook into hooks/logic directory
